Move Packet wrapping for IO to own static class
